Updated UI of Dashboard
