+++ package.json
@@ -27,9 +27,9 @@
     "@mui/material": "^5.15.4",
     "@radix-ui/react-accordion": "^1.2.3",
     "@radix-ui/react-slot": "^1.1.2",
-    "@szhsin/react-accordion": "^1.4.0",
-    "@radix-ui/react-tooltip": "^1.1.8",
     "@radix-ui/react-tabs": "^1.1.3",
+    "@radix-ui/react-tooltip": "^1.1.8",
+    "@szhsin/react-accordion": "^1.4.0",
     "bcryptjs": "^2.4.3",
     "chart.js": "^4.4.9",
     "class-variance-authority": "^0.7.1",
